% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exactGradient}
\alias{exactGradient}
\title{Gradient for the Lorenz model}
\usage{
exactGradient(X, K, M, dt, F)
}
\arguments{
\item{X}{initial state}

\item{K}{K from the Lorenz 04 model}

\item{M}{the number of steps that make a time step}

\item{dt}{time step}

\item{F}{F from the Lorenz 04 model}
}
\description{
Gradient for the Lorenz model
}
