% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz04.R
\docType{package}
\name{Lorenz04}
\alias{Lorenz04}
\alias{Lorenz04-package}
\title{Lorenz04: R implementation of models in "Designing chaotic models"}
\description{
This package can be used for testing data assimilation methods
}
